pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "block-producer-locator"
version = "0.1.0"
description = "Locate block-producing nodes in blockchain networks by IP and geographic location"
readme = "PROJECT.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "click>=8.3,<9",
    "geoip2>=5.2,<6",
    "pyyaml>=6.0,<7",
    "requests>=2.32,<3",
    "rich>=14.3,<15",
    "web3>=7.14,<8",
]

[project.optional-dependencies]
dev = [
    "pytest>=9,<10",
    "ruff>=0.15",
]

[project.scripts]
bpl = "bpl.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["bpl"]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I", "UP"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
